refactor(guard): split PreToolUse modules for the quality ratchet
