+++ .github/workflows/rust-pretool-adversarial.yml
@@ -6,6 +6,7 @@ on:
     paths:
       - "rust/**"
       - "src/codex_plugin_scanner/guard/native_command_model.py"
+      - "src/codex_plugin_scanner/guard/native_pretool.py"
       - "src/codex_plugin_scanner/guard/daemon/hook_worker.py"
       - "scripts/integration/rust_pretool_adversarial.py"
       - ".github/workflows/rust-pretool-adversarial.yml"
@@ -14,6 +15,7 @@ on:
     paths:
       - "rust/**"
       - "src/codex_plugin_scanner/guard/native_command_model.py"
+      - "src/codex_plugin_scanner/guard/native_pretool.py"
       - "src/codex_plugin_scanner/guard/daemon/hook_worker.py"
       - "scripts/integration/rust_pretool_adversarial.py"
       - ".github/workflows/rust-pretool-adversarial.yml"
